Created get_hardcode_mnist_latent_map funciton in utils.py
