+++ src/class_cvae/train_yae.py
@@ -15,7 +15,7 @@ from PIL import Image
 from models import Encoder, Decoder, Classifier, ImageClassifier, SimpleEncoder, HandCraftedMNISTDecoder
 from logger import Logger
 from lpips.lpips import LPIPS
-from utils import init_weights
+from utils import init_weights, get_hardcode_mnist_latent_map
 
 """
 Goal: Train a variational autoencoder with a classification head on the latent space.
@@ -156,18 +156,7 @@ def get_args():
     return parser.parse_args()
 
 def create_z_from_label(lbls):
-    z_map = {
-        0: np.array([[1, 0, 1, 1, 1, 1, 1]]),
-        1: np.array([[0, 0, 0, 0, 1, 0, 1]]),
-        2: np.array([[1, 1, 1, 0, 1, 1, 0]]),
-        3: np.array([[1, 1, 1, 0, 1, 0, 1]]),
-        4: np.array([[0, 1, 0, 1, 1, 0, 1]]),
-        5: np.array([[1, 0, 1, 1, 0, 0, 1]]),
-        6: np.array([[1, 1, 1, 1, 0, 1, 1]]),
-        7: np.array([[1, 0, 0, 0, 1, 0, 1]]),
-        8: np.array([[1, 1, 1, 1, 1, 1, 1]]),
-        9: np.array([[1, 1, 0, 1, 1, 0, 1]]),
-    }
+    z_map = get_hardcode_mnist_latent_map()
 
     z = z_map[lbls[0].item()]
     for i in range(1, len(lbls)):
