+++ src/class_cvae/visual_counter_factual_split.py
@@ -15,7 +15,7 @@ from PIL import Image, ImageDraw, ImageFont
 import matplotlib.pyplot as plt
 
 from models import Encoder, Decoder, Classifier, ImageClassifier
-from utils import create_img_from_text, MaxQueue, set_seed, save_tensor_as_graph
+from utils import create_img_from_text, MaxQueue, set_seed, save_tensor_as_graph, calc_img_diff_loss
 
 """
 Goal: Create visual counterfactual
@@ -143,11 +143,6 @@ def get_args():
     parser.add_argument('--num_class_features', type=int, default=7)
 
     return parser.parse_args()
-
-def calc_img_diff_loss(org_img_recon, imgs_recon, loss_fn):
-    diffs = (org_img_recon - imgs_recon)
-    loss = min_loss_fn(diffs, torch.zeros_like(diffs).cuda())
-    return loss
 
 if __name__ == "__main__":
     set_seed()
